requirements.txt
pygame==2.0.1
Pillow==8.4.0